#5 better testing, prepated library to split api and ble calls
