+++ .vscode/settings.json.DISABLED
@@ -1,9 +1,13 @@
 {
     "python.linting.pylintEnabled": true,
     "python.linting.enabled": true,
-    "python.pythonPath": "/usr/local/bin/python",
     "files.associations": {
         "*.yaml": "home-assistant"
+    },
+    "files.watcherExclude": {
+        "**/.git/objects/**": true,
+        "**/.git/subtree-cache/**": true,
+        "**/.tox/**": true
     },
     "jupyter.debugJustMyCode": false,
     "python.testing.unittestEnabled": false,
