+++ workspace.code-workspace
@@ -1,0 +1,64 @@
+{
+	"folders": [
+		{
+			"name": "hacs-govee",
+			"path": "."
+		},
+		{
+			"name": "python-govee-api",
+			"path": ".git-subtree/python-govee-api"
+		}
+	],
+	"settings": {
+		"files.associations": {
+			"*.yaml": "home-assistant"
+		},
+		"python.linting.pylintEnabled": true,
+		"python.linting.enabled": true,
+		"files.associations": {
+			"*.yaml": "home-assistant"
+		},
+		"files.watcherExclude": {
+			"**/.git/objects/**": true,
+			"**/.git/subtree-cache/**": true,
+			"**/.tox/**": true
+		},
+		"jupyter.debugJustMyCode": false,
+		"python.testing.unittestEnabled": false,
+		"python.testing.nosetestsEnabled": false,
+		"python.testing.pytestEnabled": true,
+		"python.testing.pytestArgs": [
+			"tests"
+		],
+		"docker.host": "ssh://root@192.168.144.5",
+		"docker.imageBuildContextPath": "/usr/share/hassio/share/dev/hacs-govee",
+		"launch": {
+			"configurations": [],
+			"compounds": [
+				{
+					"name": "Launch Home Assistant UI in Chrome",
+					"configurations": [
+						"Launch Home Assistant UI in Chrome"
+					]
+				},
+				{
+					"name": "Python: Remote Attach",
+					"configurations": [
+						"Python: Remote Attach"
+					]
+				},
+				{
+					"name": "Library: launch readme_example.py",
+					"configurations": [
+						"run /example/readme_example.py"
+					]
+				}
+			]
+		},
+		"tasks": {
+			"version": "2.0.0",
+			"tasks": []
+		},
+		"python.pythonPath": "/usr/local/bin/python"
+	},
+}
